# frontend/openapi.yaml
openapi: 3.1.0
info:
  title: API Documentation
  version: "1.0"
paths:
...
  /knowledge/retrieval/test:
    post:
      tags:
      - knowledge-admin-controller
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeAdminRetrievalRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultKnowledgeAdminRetrievalResponse"
  /knowledge/files:
    get:
      tags:
      - knowledge-admin-controller
      parameters:
      - name: kbId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          default: desc
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListKnowledgeFileVO"
    post:
      tags:
      - knowledge-admin-controller
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                kbId:
                  type: integer
                  format: int64
                file:
                  type: string
                  format: binary
        required: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultDocumentIngestionResult"
  /knowledge/files/{id}/reingest:
    post:
      tags:
      - knowledge-admin-controller
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultDocumentIngestionResult"
...
  /knowledge/files/{id}:
    get:
      tags:
      - knowledge-admin-controller
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultKnowledgeFileVO"
    delete:
      tags:
      - knowledge-admin-controller
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultBoolean"
...
components:
  schemas:
...
    ResultBoolean:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          type: boolean
...
    KnowledgeAdminRetrievalRequest:
      type: object
      properties:
        kbId:
          type: integer
          format: int64
        query:
          type: string
        topK:
          type: integer
          format: int32
        useVector:
          type: boolean
    Candidate:
      type: object
      properties:
        chunkId:
          type: string
        text:
          type: string
        score:
          type: number
          format: double
        metadata:
          type: object
          additionalProperties: {}
        knowledgeBaseId:
          type: integer
          format: int64
        filename:
          type: string
    KnowledgeAdminRetrievalResponse:
      type: object
      properties:
        query:
          type: string
        topK:
          type: integer
          format: int32
        usedVector:
          type: boolean
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
    ResultKnowledgeAdminRetrievalResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/KnowledgeAdminRetrievalResponse"
    DocumentIngestionResult:
      type: object
      properties:
        uploadedFileId:
          type: integer
          format: int64
        documentId:
          type: string
        chunkCount:
          type: integer
          format: int32
        vectorCount:
          type: integer
          format: int32
    ResultDocumentIngestionResult:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/DocumentIngestionResult"
...
    KnowledgeFileVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: string
        originalFilename:
          type: string
        fileType:
          type: string
        status:
          type: string
        segmentCount:
          type: integer
          format: int32
        vectorCount:
          type: integer
          format: int32
        knowledgeBaseId:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        errorMessage:
          type: string
    ResultListKnowledgeFileVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeFileVO"
    ResultKnowledgeFileVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/KnowledgeFileVO"
...